redid completely; now creates and redirects
This is no longer a place for users to edit documents. This page will give the user a token which is used to create a document. Then, the user is redirected to their document and the token is deleted. A second document will not be made after because the token was deleted; if the user goes back to that page they will see "session expired"