ctx-menu, modals, refactored card loading
card loading now loads them async one by one. there is a mechanism in place to prevent users from spamming reload and wasting my bandwidth, or users leaving to the next page or reloading just to go back and have to wait for more. The docs are saved locally (for like 15 seconds) in instances like these to quickly repopulate /documents/